Updated OLM CRDs and catalog-operator template
The OLM project recently switched to using Cobra for CLI flag management. This changed the command arguments for the catalog container to be double hyphenated '--' instead of single '-'.

We use the master/latest tag for the operator-framework/olm image, which was no longer compatible with our helm chart configuration. This update brings in those changes, along with a minor update to the CRDs